moved followListens, myFeed, similarListens pagingSources to shared module
