+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/ui/screens/feed/MyFeedPagingSource.kt
@@ -1,22 +1,26 @@
-package org.listenbrainz.android.ui.screens.feed
+package org.listenbrainz.shared.ui.screens.feed
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import org.listenbrainz.shared.model.ResponseError
-import org.listenbrainz.android.repository.feed.FeedRepository
+import org.listenbrainz.shared.model.feed.FeedData
+import org.listenbrainz.shared.model.feed.FeedEventType
+import org.listenbrainz.shared.repository.feed.FeedRepository
 import org.listenbrainz.shared.util.Resource
+import kotlin.time.Clock
 
-class SimilarListensPagingSource(
+class MyFeedPagingSource (
     private val username: suspend () -> String,
+    private val addEntryToMap: (Int, Boolean) -> Unit,
     private val onError: (error: ResponseError?) -> Unit,
     private val feedRepository: FeedRepository,
     private val ioDispatcher: CoroutineDispatcher
 ): PagingSource<Long, FeedUiEventItem>() {
-    
-    override fun getRefreshKey(state: PagingState<Long, FeedUiEventItem>): Long? {
-        return System.currentTimeMillis() / 1000
+
+    override fun getRefreshKey(state: PagingState<Long, FeedUiEventItem>): Long {
+        return Clock.System.now().epochSeconds
     }
     
     override suspend fun load(params: LoadParams<Long>): LoadResult<Long, FeedUiEventItem> {
@@ -29,13 +33,14 @@ class SimilarListensPagingSource(
         }
         
         val result = withContext(ioDispatcher) {
-            feedRepository.getFeedSimilarListens(username = username, maxTs = params.key, count = params.loadSize)
+            feedRepository.getFeedEvents(username = username, maxTs = params.key, count = params.loadSize)
         }
         
         return when (result.status) {
             Resource.Status.SUCCESS -> {
                 
-                val processedEvents = FollowListensPagingSource.processFeedEvents(result.data)
+                val processedEvents = processFeedEvents(result.data)
+
                 val nextKey = processedEvents.lastOrNull()?.event?.created?.let { newKey ->
                     // Termination condition.
                     if (params.key != null && newKey >= params.key!!)
@@ -46,9 +51,9 @@ class SimilarListensPagingSource(
                         )
                     else
                         newKey
-        
+                    
                 }
-                
+
                 LoadResult.Page(
                     data = processedEvents,
                     prevKey = null,
@@ -63,4 +68,27 @@ class SimilarListensPagingSource(
         }
         
     }
+    
+    private fun processFeedEvents(feedData: FeedData?): List<FeedUiEventItem> {
+        
+        return mutableListOf<FeedUiEventItem>().apply {
+            
+            feedData?.payload?.events?.forEach { event ->
+                
+                // Add the entry to map.
+                if (event.hidden == true) {
+                    event.id?.let { addEntryToMap(it, true) }
+                }
+                
+                add(
+                    FeedUiEventItem(
+                        event = event,
+                        eventType = FeedEventType.resolveEvent(event),
+                        parentUser = feedData.payload.userId
+                    )
+                )
+            }
+        }
+    }
+    
 }
