+++ shared/src/commonMain/kotlin/org/listenbrainz/shared/ui/screens/feed/SimilarListensPagingSource.kt
@@ -1,24 +1,23 @@
-package org.listenbrainz.android.ui.screens.feed
+package org.listenbrainz.shared.ui.screens.feed
 
 import androidx.paging.PagingSource
 import androidx.paging.PagingState
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import org.listenbrainz.shared.model.ResponseError
-import org.listenbrainz.android.model.feed.FeedData
-import org.listenbrainz.android.model.feed.FeedEventType
-import org.listenbrainz.android.repository.feed.FeedRepository
+import org.listenbrainz.shared.repository.feed.FeedRepository
 import org.listenbrainz.shared.util.Resource
+import kotlin.time.Clock
 
-class FollowListensPagingSource(
+class SimilarListensPagingSource(
     private val username: suspend () -> String,
     private val onError: (error: ResponseError?) -> Unit,
     private val feedRepository: FeedRepository,
     private val ioDispatcher: CoroutineDispatcher
 ): PagingSource<Long, FeedUiEventItem>() {
     
-    override fun getRefreshKey(state: PagingState<Long, FeedUiEventItem>): Long {
-        return (System.currentTimeMillis() / 1000)
+    override fun getRefreshKey(state: PagingState<Long, FeedUiEventItem>): Long? {
+        return Clock.System.now().epochSeconds
     }
     
     override suspend fun load(params: LoadParams<Long>): LoadResult<Long, FeedUiEventItem> {
@@ -31,12 +30,13 @@ class FollowListensPagingSource(
         }
         
         val result = withContext(ioDispatcher) {
-            feedRepository.getFeedFollowListens(username = username, maxTs = params.key, count = params.loadSize)
+            feedRepository.getFeedSimilarListens(username = username, maxTs = params.key, count = params.loadSize)
         }
         
         return when (result.status) {
             Resource.Status.SUCCESS -> {
-                val processedEvents = processFeedEvents(result.data)
+                
+                val processedEvents = FollowListensPagingSource.processFeedEvents(result.data)
                 val nextKey = processedEvents.lastOrNull()?.event?.created?.let { newKey ->
                     // Termination condition.
                     if (params.key != null && newKey >= params.key!!)
@@ -47,6 +47,7 @@ class FollowListensPagingSource(
                         )
                     else
                         newKey
+        
                 }
                 
                 LoadResult.Page(
@@ -59,28 +60,8 @@ class FollowListensPagingSource(
                 onError(result.error)
                 LoadResult.Error(Exception(result.error?.toast))
             }
-        }
-    }
-    
-    companion object {
-        
-        /** Converts [feedData] to list of [FeedUiEventItem] for presentation.*/
-        fun processFeedEvents(feedData: FeedData?): List<FeedUiEventItem> {
-        
-            return mutableListOf<FeedUiEventItem>().apply {
             
-                feedData?.payload?.events?.forEach { event ->
-                    add(
-                        FeedUiEventItem(
-                            event = event,
-                            eventType = FeedEventType.resolveEvent(event),
-                            parentUser = feedData.payload.userId
-                        )
-                    )
-                }
-            }
         }
+        
     }
-    
-    
 }
